+++ index.js
@@ -26,10 +26,15 @@ const Invoices        = require('./routes/Invoice');
 const Application     = require('./routes/Application');
 const Developer       = require('./routes/Developer');
 
+const Common = require('./common/Common');
+
 start = function() {
 
   // Initialize the program
   program.version(pkg.version);
+
+  // Check for available new version
+  Common.checkNewVersion();
 
   let prefs = new Preferences();
 
@@ -75,6 +80,8 @@ start = function() {
         Message.lineFeed();
         Message.notFound();
   });
+
+
 
   program.parse(process.argv);
 }